requirements.txt
supabase>=2.0.0
python-dotenv>=1.0.0
pydantic>=2.0.0
pandas>=2.0.0
rapidfuzz>=3.0.0
tqdm>=4.65.0
pdfplumber>=0.11.0
requests>=2.31.0
beautifulsoup4>=4.12.0
